systemd/timeinit: add HTTPS time synchronisation service
Add a new timesync-https systemd service to synchronise the system
time at boot using an HTTPS header. The service uses curl to request
an HTTPS header from either $API_ENDPOINT/connectivity-check (default)
or the URL defined by the os.network.connectivity.uri field in
config.json. The URL used *must* return HTTP code 204 (No Content)
in response to a request so that we can determine that we have full
network connectivity and are not operating behind a captive portal.

The date field returned by a valid header is used to set the system
time providing that it is in advance of the current system time.

The service will exit when a valid response has been received. Poll
attempts will be made at an increasing interval starting at 2s and
doubling up to a maximum of 64s. Polling will continue at the maximum
interval until a valid response has been received.

The service provides time synchronisation for devices where NTP is
blocked. For devices where NTP is available it should ensure that any
system 'time jump' is only a few seconds when NTP synchronisation is
eventually achieved. It also allows other services to start with a
reasonably accurate time without having to wait for the NTP
synchronisation process to complete.

Services that are ordered after the new time-sync-https-wait target
can be sure that full network connectivity has been achieved and that
time has been synchronised with an accuracy of a few seconds.

Change-type: minor
Connects-to: #1337 #1776 #2044 #2139
Signed-off-by: Mark Corbin <mark@balena.io>